starlark_lsp::server

Trait LspContext

pub trait LspContext {
    // Required methods
    fn parse_file_with_contents(
        &self,
        uri: &LspUrl,
        content: String,
    ) -> LspEvalResult;
    fn resolve_load(
        &self,
        path: &str,
        current_file: &LspUrl,
        workspace_root: Option<&Path>,
    ) -> Result<LspUrl>;
    fn render_as_load(
        &self,
        target: &LspUrl,
        current_file: &LspUrl,
        workspace_root: Option<&Path>,
    ) -> Result<String>;
    fn resolve_string_literal(
        &self,
        literal: &str,
        current_file: &LspUrl,
        workspace_root: Option<&Path>,
    ) -> Result<Option<StringLiteralResult>>;
    fn get_load_contents(&self, uri: &LspUrl) -> Result<Option<String>>;
    fn get_environment(&self, uri: &LspUrl) -> DocModule;
    fn get_url_for_global_symbol(
        &self,
        current_file: &LspUrl,
        symbol: &str,
    ) -> Result<Option<LspUrl>>;

    // Provided methods
    fn parse_file(&self, uri: &LspUrl) -> Result<Option<LspEvalResult>> { ... }
    fn get_string_completion_options(
        &self,
        document_uri: &LspUrl,
        kind: StringCompletionType,
        current_value: &str,
        workspace_root: Option<&Path>,
    ) -> Result<Vec<StringCompletionResult>> { ... }
}

Various pieces of context to allow the LSP to interact with starlark parsers, etc.

Required Methods

fn parse_file_with_contents( &self, uri: &LspUrl, content: String, ) -> LspEvalResult

Parse a file with the given contents. The filename is used in the diagnostics.

fn resolve_load( &self, path: &str, current_file: &LspUrl, workspace_root: Option<&Path>, ) -> Result<LspUrl>

Resolve a path given in a load() statement.

path is the string representation in the load() statement. Its meaning is implementation defined. current_file is the the file that is including the load() statement, and should be used if path is “relative” in a semantic sense.

fn render_as_load( &self, target: &LspUrl, current_file: &LspUrl, workspace_root: Option<&Path>, ) -> Result<String>

Render the target URL to use as a path in a load() statement. If target is in the same package as current_file, the result is a relative path.

target is the file that should be loaded by load(). current_file is the file that the load() statement will be inserted into.

fn resolve_string_literal( &self, literal: &str, current_file: &LspUrl, workspace_root: Option<&Path>, ) -> Result<Option<StringLiteralResult>>

Resolve a string literal into a Url and a function that specifies a location within that target file.

This can be used for things like file paths in string literals, build targets, etc.

current_file is the file that is currently being evaluated

fn get_load_contents(&self, uri: &LspUrl) -> Result<Option<String>>

Get the contents of a starlark program at a given path, if it exists.

fn get_environment(&self, uri: &LspUrl) -> DocModule

Get the preloaded environment for a particular file.

fn get_url_for_global_symbol( &self, current_file: &LspUrl, symbol: &str, ) -> Result<Option<LspUrl>>

Get the LSPUrl for a global symbol if possible.

The current file is provided in case different files have different global symbols defined.

Provided Methods

fn parse_file(&self, uri: &LspUrl) -> Result<Option<LspEvalResult>>

Get the contents of a file at a given URI, and attempt to parse it.

fn get_string_completion_options( &self, document_uri: &LspUrl, kind: StringCompletionType, current_value: &str, workspace_root: Option<&Path>, ) -> Result<Vec<StringCompletionResult>>

Get valid completion options if possible, based on the kind of string completion expected (e.g. any string literal, versus the path argument in a load statement).